test(discord): extract shared consumed-flip test helpers
Round 8 of PR review on #835. Test-only DRY cleanup.

cr round-8 #4: flipVerdict/flipVerdictLog + drainTicks (and flushFlip)
were reimplemented across qurl-webhook-consumed.test.js and
qurl-webhook-consumed-payload-throw.test.js. Extracted them into
tests/helpers/consumed-flip.js (per the repo's tests/helpers/
convention), parameterized by the caller's mocked logger. Both suites
now bind thin wrappers over the shared helper; call sites unchanged.

cr round-8 #1 (comment density), #2 (status-string surface shrank — true,
external HTTP behavior unchanged), #3 (shutdown gap) are repeats already
acknowledged / no-change.

83 suites / 2804 tests, lint green; qurl-webhook.js coverage 96.23%.
